+++ src/main/java/ca/corefacility/bioinformatics/irida/ria/web/projects/ProjectSamplesController.java
@@ -5,7 +5,15 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.attribute.BasicFileAttributes;
 import java.security.Principal;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Date;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Locale;
+import java.util.Map;
+import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipOutputStream;
@@ -43,7 +51,6 @@ import com.github.dandelion.datatables.core.export.ExportUtils;
 import com.github.dandelion.datatables.core.export.ReservedFormat;
 import com.github.dandelion.datatables.extras.spring3.ajax.DatatablesParams;
 import com.google.common.base.Strings;
-import com.google.common.collect.ImmutableMap;
 
 import ca.corefacility.bioinformatics.irida.exceptions.EntityExistsException;
 import ca.corefacility.bioinformatics.irida.exceptions.EntityNotFoundException;
@@ -369,7 +376,6 @@ public class ProjectSamplesController {
 			@RequestParam(value = "sampleNames[]") List<String> sampleNames, Locale locale) {
 		// Need to keep the count for comparison after.
 		int originalCount = sampleNames.size();
-		List<Join<Project, Sample>> samples = new ArrayList<>();
 
 		// Get a list of all samples for all projects
 		projects.add(0, projectId);
@@ -705,7 +711,8 @@ public class ProjectSamplesController {
 
 		if (!Strings.isNullOrEmpty(newName)) {
 			try {
-				mergeIntoSample = sampleService.update(mergeSampleId, ImmutableMap.of("sampleName", newName));
+				mergeIntoSample.setSampleName(newName);
+				mergeIntoSample = sampleService.update(mergeIntoSample);
 			} catch (ConstraintViolationException e) {
 				logger.error(e.getLocalizedMessage());
 				result.put("result", "error");
@@ -947,7 +954,6 @@ public class ProjectSamplesController {
 
 		Project project = projectService.read(projectId);
 		List<Project> projects = new ArrayList<>();
-		List<Sample> samples = new ArrayList<>();
 
 		if (!associated.isEmpty()) {
 			projects = (List<Project>) projectService.readMultiple(associated);